Changing the price datatype to long and refractored the code

+++ src/main/java/com/maha/discounts/strategy/DiscountStrategy.java
@@ -8,5 +8,5 @@ import com.maha.model.discounts.ProductDiscount;
  */
 public interface DiscountStrategy {
 
-    double apply(ProductOrder product, ProductDiscount productDiscount);
+    long apply(ProductOrder product, ProductDiscount productDiscount);
 }
